fix: address issue:4644047856 — split test_pr_monitor_part_005 to restore 1500-line guardrail
Subsequent commits pushed test_pr_monitor_part_005.py from 1472 lines (compliant
at Greptile review time) to 1540 lines, breaching the enforced 1500-line cap
checked by test_first_party_code_files_stay_under_line_limit.

Move TestImageRefMatchesDaemonUrl into a new test_pr_monitor_part_006.py (662 lines)
so part_005 returns to 935 lines. The split duplicates only the _status() helper;
no test logic is changed.
